+++ src/model/Country.java
@@ -3,11 +3,11 @@ package model;
 public class Country {
     private String id; //UUID
     private String name;
-    private int population;
+    private long population;
     private String countryCode;
 
     // Constructor
-    public Country(String id, String name, int population, String countryCode) {
+    public Country(String id, String name, long population, String countryCode) {
         this.id = id; // 6ec3e8ec-3dd0-11ed-b878-0242ac120002
         this.name = name; // Colombia
         this.population = population; // 50.2
@@ -22,8 +22,8 @@ public class Country {
     public String getName() {return name;}
     public void setName(String name) {this.name = name;}
 
-    public int getPopulation() {return population;}
-    public void setPopulation(int population) {this.population = population;}
+    public long getPopulation() {return population;}
+    public void setPopulation(long population) {this.population = population;}
 
     public String getCountryCode() {return countryCode;}
     public void setCountryCode(String countryCode) {this.countryCode = countryCode;}
